+++ src/renderer/src/modules/albums/albums-screen/index.ts
@@ -1,1 +1,0 @@
-export { default as AlbumsScreen } from './albums-screen';