+++ src/renderer/src/modules/albums/albums-sidebar/index.ts
@@ -1,1 +1,0 @@
-export { default as AlbumSidebar } from './albums-sidebar';