+++ src/renderer/src/modules/photos/photos-screen/index.ts
@@ -1,1 +1,0 @@
-export { default as PhotosScreen } from './photos-screen';