+++ src/renderer/src/modules/photos/photos-screen/photos-screen.tsx
@@ -1,5 +1,0 @@
-const PhotosScreen = () => {
-  return <div className="flex">Photos</div>;
-};
-
-export default PhotosScreen;